Rename channel manager to channel resolver
