+++ src/channel_resolver_ext.rs
